requirements.txt
azure-core==1.33.0
azure-cosmos==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
requests==2.32.3
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
